refactor: move logic for saving records inside SalesViewModel
